Docker_lambda 2/requirements.txt
openai 
pypdf
pinecone-client 
langchain 
langchain-community
tiktoken
Requests
